+++ packages/infinity/src/modules.js
@@ -8,6 +8,7 @@ import RecipeManagement from '@shopworx/recipe-management';
 import ParameterConfiguration from '@shopworx/parameter-configuration';
 import RoadmapManagement from '@shopworx/roadmap-management';
 import ProductionLayout from '@shopworx/production-layout';
+import MaterialManagement from '@shopworx/material-management';
 import ProductManagement from '@shopworx/product-management';
 import ReportViewer from '@shopworx/report-viewer';
 import Setup from '@shopworx/setup';
@@ -30,6 +31,12 @@ Vue.use(RoadmapManagement, {
 });
 
 Vue.use(ParameterConfiguration, {
+  router,
+  store,
+  i18n,
+});
+
+Vue.use(MaterialManagement, {
   router,
   store,
   i18n,